:sparkles: feat: add EventForm component
