+++ src/pages/create.js
@@ -1,22 +1,17 @@
-import { FormProvider, useForm } from "react-hook-form";
-
 import { ArrowBackRounded } from "@mui/icons-material";
-import { AppBar, Box, Button, IconButton, Stack, Toolbar, Typography } from "@mui/material";
+import { AppBar, Box, IconButton, Toolbar, Typography } from "@mui/material";
 
 import Link from "next/link";
 import { useRouter } from "next/router";
 
 import { useMutation } from "@tanstack/react-query";
 
-import { Input } from "../components/Input";
-
 import { createEvent } from "../api";
 
 import { removeSeconds } from "../ultis/date";
+import { EventForm } from "../components/EventForm";
 
 export default function Create() {
-  const methods = useForm();
-
   const { push } = useRouter();
 
   const createEventMutation = useMutation(createEvent);
@@ -50,21 +45,7 @@ export default function Create() {
         <Typography color="text.secondary">Adicione acessos, palestras ou peças</Typography>
 
         <Box mt={1}>
-          <FormProvider {...methods}>
-            <form onSubmit={methods.handleSubmit(handleSave)}>
-              <Input name="title" label="Nome do evento" />
-
-              <Input.Time name="startDate" label="Início" defaultValue={null} />
-
-              <Input.Time name="endDate" label="Fim" defaultValue={null} />
-
-              <Stack direction="row" justifyContent="flex-end">
-                <Button sx={{ mt: 4 }} type="submit" variant="contained">
-                  Salvar
-                </Button>
-              </Stack>
-            </form>
-          </FormProvider>
+          <EventForm onSubmit={handleSave} />
         </Box>
       </Box>
     </Box>
